rename Drivers dir to snake case
