+++ CMakeLists.txt
@@ -26,10 +26,10 @@ set(C2A_CORE_DIR ${CMAKE_CURRENT_SOURCE_DIR})
 
 set(C2A_SRCS
   c2a_core_main.c
-  Drivers/common_tlm_cmd_packet_for_driver_super.c
-  Drivers/eb90_frame_for_driver_super.c
-  Drivers/eb90_packet_for_driver_super.c
-  Drivers/driver_super.c
+  drivers/common_tlm_cmd_packet_for_driver_super.c
+  drivers/eb90_frame_for_driver_super.c
+  drivers/eb90_packet_for_driver_super.c
+  drivers/driver_super.c
   hal/i2c_common.c
   system/application_manager/app_info.c
   system/application_manager/app_manager.c
