+++ script/ci/check_encoding.json
@@ -32,7 +32,7 @@
   },
   "target_dirs" : [
     "applications/",
-    "Drivers/",
+    "drivers/",
     "hal/",
     "library/",
     "System/",
